Fix race condition in MasterKeys.getOrCreate
This was reported to me back in December, 2022 but has been
difficult to reproduce. I've only seen it occur in Tink
integration tests hosted on an internal google repo. Attempts to
reproduce in my Android Studio environment have failed, but the
included patch fixed the race condition in the one environment
it could be demonstrated in.

Relnote: Fixed a race condition in MasterKeys.getOrCreate
Test: See above
Bug: 268572037

Change-Id: I3391e7ba62acfa942432eba402e74d64e2e5e98a