Fix proguard parse error
It seems like the existing proguard config had a syntax error even
though R8 did not complain about it. This seems to cause an issue when
DexGuard is used.

Tested by analysing the APK and verifying that sProcessGlobalConfig is retained in both cases (with * and with ***) by R8. That said, according to https://playground.proguard.com/p/bmaZvZ, it seems like '*' is not accepted syntax even though it is accepted by R8. '***' is accepted syntax.

Bug: 270034835
Test: ./gradlew webkit:integration-tests:testapp:connectedAndroidTest
Change-Id: Ia65c2fd50327eda385bc0ec061f4ad08a7afe81d