Handle Dynamically Updating/Visible Pointer Icons
Updates the pointer icon logic to additionally include use cases that
address the pointer icon dynamically changing while the user is hovered
over the element, as well as handling the use case where a parent/child
in the hierarchy of custom pointer icons is dynamically added/removed.

Test: Added tests for icons dynamically updating mid-test, as well as
test cases for dynamically removing a parent, child, and
child/grandchild combo.
Bug:269615253

Change-Id: Ia2e7bea276b0dd821a78a6a27fcad76dc4982f22