Define ChangeList to manage pending changes
This CL is the first in a multi-step refactor of how operations are
tracked by the Composer. Currently, we allocate a lambda for each
action we want to execute when running the changes, which can lead to
alloc/GC churn. This change introduces an `OpStack` type, which can
track a bespoke set of `Operation`s and their corresponding arguments.

We utilize three arrays here (one for the operations, one for `int`
arguments, and one for all other arguments), which helps us reduce
allocations and prevent autoboxing.

As an added bonus, the ChangeList can be formatted as a string to
reveal what operations are pending, which can be a big help for
debugging.

This CL just attempts to introduce the data structure changes. As an
intermediate step, we're using a `BackwardsCompatOp` that takes in a
lambda and behaves exactly the same as the previous behavior of
allocating lambdas. This operation will be removed by the end of the
refactor sequence.

Test: OpStackTest, ./gradlew :compose:runtime:runtime:tDUT
Change-Id: I70eabbddb874c24c926a05ea624f566ff878922a