Always call enter and exit focus properties
We currently call enter and exit focus properties
during focus search, but not if we manually call
requestFocus or clearFocus. Calling these properties
manually has led to bugs in the past (b/b/267827550),
and this helps clean up the code too.

It should also be helpful for usecases where a parent
wants to intercept clear focus calls, and other
interop use cases.

Bug: 267827550
Test: Existing tests pass and also added new tests.
Change-Id: I5898fef5a7dafea2ecc8aa1ec4131226d31540dd