Adds API for setting dates at the DatePicker state
This change adds functionality to the DatePickerState and the
DateRangePickerState to allow setting dates after the initial creation
of the state, as well as reset the selections by setting null values.

Note that setting dates should be make with caution. The new
`setSelection` function will throw an exception in case the provided
dates fail some preconditions (see documentation).

Fixes: 268609314
Fixes: 270427389
Test: DatePickerTest, DateRangePickerTest
Relnote: "Added functionality to the DatePickerState and the
DateRangePickerState to allow setting dates after the initial creation
of the state, as well as resetting selections by setting null
timestamps. Setting dates manually should be made with caution, and
the new `setSelection` function will throw an exception in case a
provided date fails a precondition (see documentation)."

Change-Id: Ifa6451a9e32ad031ac3223cb8b5fea132a2d0191