Apply ResolutionCorrector onto the output sizes list before using it to calculate the suggested stream spec
The ResolutionCorrector was incorrectly removed in aosp/2452267. Most of the output sizes corrector logic code has been moved to the SupportedOutputSizesSorter class in camera-core. The ResolutionCorrector workaround can't be moved to the core module. Instead, we apply the workaround before using the sorted output sizes to calculate the suggested stream spec.

Bug: 270825043
Test: SupportedSurfaceCombinationTest
Change-Id: I40df52b7fe707839b683500f2512129772a2ab52