Refactoring graphics group to allow sub-modules that ship independently
The GRAPHICS library group was set up to be atomic. Then we started
adding sub-groups (eg, graphics-shapes) that have nothing to do with
graphics-core and should not ship at the same time.

This change keeps GRAPHICS as a group, but makes it non-atomic, and
adds GRAPHICS_CORE and GRAPHICS_SHAPES itmes in LibraryVersions.toml
to enable individual release versions for these separate libraries.

Bug: N/A
Test: Built/ran the shapes library
Merged-In: Ic152102b26ea7ab2f59628ee62003e6962d079e0
Change-Id: Ic152102b26ea7ab2f59628ee62003e6962d079e0