Allow an empty VirutalCameraCaptureResult with the timestamp overriden.
This is part of the work that enables extensions on the new ImageCapture pipeline.

When a JPEG Surface is not connected to the camera, e.g. for extentions or stream sharing, the ImageReader will not receive a CameraCaptureResult with matching timestamp. As a workaround, we need to create a VirutalCameraCaptureResult in CaptureNode with only the timestamp of the Image.

Bug: 271299711
Test: manual test and ./gradlew bOS
Change-Id: I788bebe322b7aa86f8328a1f5049b31dc0bf9d5f