Add an option for exporting schemas as JAR resources.
The 'room.exportSchemaResource' will make it such that Room will use Filer APIs to write the output schema as a resources that will end up in the produced artifact. This is necessary for build systems that strictly prohibit annotation processor from writing to arbitrary locations (i.e. using io.File APIs) instead of using the processing environment Filer.

Note that schema files are never intended to be left in the producer JAR, thus this option should only be used once the user is ready to check-in the schema file into the repository as described in https://developer.android.com/training/data-storage/room/migrating-db-versions. To prevent accidents, enabling the option will cause Room to produce a warning hinting to its intended usage.

Bug: 265412779
Test: DatabaseProcessorTest
Change-Id: Ie3f879b23dd612929ea0c4995840a1288a19134f