Introduced per-variant configuration
This PR introduces per variant configurations. All the flags can now
be expressed per variant. For example:
```
baselineProfile {

    // Applied to all the variants
    saveInSrc = true

    variants {

        // Applied to all the release build types
        release {

            // Overrides global config
            saveInSrc = false
        }

        // Applied to free release variant
        freeRelease {

            // Overrides any other config
            saveInSrc = true
        }
    }
}
```
Variant configurations are evaluated in with the following priority:
variant name, flavor or build type, global. Note that it's not possible
to express together a flavor and build type configuration because it'd
create ambiguity on which value to use.

For example the following configuration would generate a failure, as
it expresses a configuration for both flavor and build type.
```
baselineProfile {
    variants {
        free { ... }
        release { ... }
    }
}
```

Bug: 266214654
Test: ./gradlew :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration :benchmark:benchmark-baseline-profile-gradle-plugin:test
Change-Id: I018ffd3b273bba80d7d8d43afce59db489bfe908