Define preambles for ChangeList operations
This CL moves more operation logic out of the Composer and into
ChangeList. Previously, the Composer had different logic for recording
"applier", "slot editing", and "slot table" operations. This CL moves
the logic that differentiates those operations into ChangeList.

The approach here is that operations can have preambles of required
checks that can push changes to the operation stack before adding
the explicitly desired operation. Preambles are somewhat of a
replacement for "realizing" operations, since a good amount of
realization happens inside of preambles.

To indicate that an operation has a preamble, the Operation class may
implement one (or more, technically) of the `ApplierOperation`,
`SlotEditingOperation`, and `SlotTableOperation` sealed interfaces.

Test: ./gradlew :compose:runtime:runtime:tDUT
Change-Id: I61f29e896ab88de9476fcf0bc9a18629cdd3b3fa