Make navController by lazy in NavHostFragment
Once Fragments are driven by a LifecycleObserver, the Fragment will no
longer be created in Activity `onCreate()` so attempting to retrieve a
navController would fail. We should make it lazy so it does not fail.

RelNote: "`NavHostFragment` now allows you to retrieve the
`NavController` as soon as the `NavHostFragment` is attached to the
`FragmentManager`, rather than only after `onCreate()`."
Test: all existing tests pass
Bug: 220186282

Change-Id: Ic638294d888c7c2002b6107d248d2d423a5614ee