Significantly improve the performance of SupportMenuInflater
By preventing the dispatch of change events until all items are inflated,
we only send one change event rather than n.  This can potentially be
multiple frames's worth of time with realistic numbers of menu items.

Fixes: 253460617
Test: Manual, used the sample app at
https://github.com/borysboykogoogle/ToolbarMenuInflationLatencyExploration
with an additional 500ms delay to the Fragment's menu inflation.
Confirmed that without the changes, menu inflation is much slower in
the Fragment version, and with the changes, they're about equal.

Also added new tests to MenuBuilderTest

Change-Id: I0b0875546d4817576a9d360a890eccce60555884