Prevent crash when adding unmanaged fragment in nav
Currently, if you attempt to add a fragment to the fragment manager
associated with the FragmentNavigator, navigation crashes as it cannot
find an entry associated with fragment.

While it is strongly recommended to only use navigation APIs to
manipulate the fragment manager, if the back stack is not affected, then
we should not crash.

RelNote: "Adding fragments to the `FragmentNavigator`'s
`FragmentManager` without using navigation will no longer crash."
Test: Added FragmentNavigatorTest
Bug: 274167493

Change-Id: I47d4c75ca46ee7f8282b2a7b97d28d71edd9a07b