Swipe to Reveal
Adding new component, Swipe to reveal in Wear Compose.
The component works by adding a `Modifier.swipeable` on
the composable which is displayed initially and is
configured based on the provided actions, anchors and thresholds.

Bug: 271986491
Relnote: "Swipe to Reveal for Wear Compose. It can be used to add other actions,
apart from click on the component and can be used by user to perform those actions
and accomplish other actions."
Test: N/A

Change-Id: I60862f2a9bcb904b0281e3c85b56aaf2a057a976