Introduce Material 3 Button to Compose for Wear OS
The default Button has the 'filled' background colors to be consistent with Compose Material3. Other convenient shortcuts such as FilledIconButton, OutlinedButton and ChildButton are also provided.

Test: androidx.wear.compose.material tests
Bug: 275717921
Relnote: "We have added the Material 3 Button component - formerly named Chip in the Material library (renamed to Button for consistency with Compose Material 3). The default Button has a filled background and there are button variations for FilledTonal (muted background), Outlined (transparent with a thin border) and Child (transparent background and no border, used for supplementary actions with the lowest amount of prominence)."
Change-Id: Ia69423e823d0b307a1b0a4782dcdd58659731c77