Small API adjustments to Polygon, RoundedPolygon, and CornerRounding
The distinction in the API between Polygon (superclass) and
RoundedPolygon (subclass) was subtle and confusing. In particular,
all of the functionality for rounding was in Polygon anyway, and the
difference between rounded and unrounded polygons was a bit academic
for the extra layer it added to the
API.

The new API simply has one class: RoundedPolygon. Callers can, of
course, create an unrounded polygon (as before). But at least the
name of the class makes it clearer that one can (if desired) create
a rounded version of a polygon (or not).

CornerRounding takes a radius parameter which was improperly annotated
to have a max value of 1. This works for canonical shapes with bounds
of (-1,1), but does not work if the caller supplies a shape in screen
(or other) coordinates. This fix removes that annotation and updates
the docs to be more correct.

Bug: 276469686
Bug: 276469361
Test: All automated tests pass, integration app builds/runs
RelNote: Polygon and RoundedPolygon APIs are merged into just RoundedPolygon
Change-Id: I770054fda707ba80c31b1290dbd14285ec34db25