Fix an issue with removing fragments that were not attached
When a fragment is added and removed in the same frame via a `navigate`
and `pop`, we move the Lifecycle state of entries but since the fragment
is optimized out of the fragment manager and never gets attached, the
entries do not get a signal that allows them to markTransitionComplete
and get to the proper Lifecycle state.

We should do something similar to what we do for fragments with saved
state that is cleared and used a ViewModel being cleared as the signal
to move all transitioning entries to their proper final lifecycle state.

RelNote: "When there is a `navigate` that adds an entry and a
`popBackStack` that removes it in the same frame, the resulting top
entry on the back stack will now consistently make it back to the
`RESUMED` `Lifecycle.State`."
Bug: 276495952
Test: added new tests

Change-Id: Id80678cf8856ca2de55c8adbc19b33fa0f8fbec9