Implement LazyPagingItem itemContenType as helper function
Added itemContentType API which returns a lambda that can be plugged into `contentType` params for Lazy scopes.
If no contentType is provided, defaults to null contentType.

Test: Compose demos
Test: ./gradlew paging:paging-compose:cC
Bug: 259385813
Relnote: "LazyPagingItems has a new itemContentType API which returns a contentType implementation that can be plugged into the `contentType` params of Lazy scopes. If no contentType is provided, defaults to `null` contentType for all paging items."
Change-Id: Ib04f099111605af368c5b025d84a085ab4251fb6