Allow GraphProcessor to submit parameters before the graph starts
Currently 3A functions such as lock3A() or unlock3A() will fail
immediately if we don't already have a started CameraGraph AND a
a repeating request that has already been submitted.

This CL allows GraphProcessor to still submit parameters by storing the
last parameters to submit, and retries the submission when we have a
current repeating request.

Bug: 263211462
Test: GraphProcessorTest
Change-Id: I0b6b3670daef4f4c3d20b658430a74e7ab1e21fb