Replace MutableTextFieldValueWithSelection.resetTo with revertAllChanges.
resetTo was weird: it was basically a replaceAll function, but given its
purpose in filters, it needs to also clear the changelist. However,
it accepted an arbitrary value to reset to when the only _valid_ value
was the original value passed into the filter.

The new function doesn't take any parameters, instead will just revert
to whatever the original state was passed into the constructor, and it
will also clear the changelist.

Test: existing tests cover
Relnote: n/a
Change-Id: Id2ee83d78b301b1d18104c6f18a6d40fd1430e11