Minor API adjustments
RoundedPolygon had redundant constructors; two were removed.

Star used to take an innerRadiusRatio, in addition to the outer
radius. It now takes two radius values, radius1 and radius2.
These values are absolute (neither is a ratio of the other) and
can be either the inner or outer radius (or equal, which results in
a regular polygon).

Bug: 277732981
Test: Built/ran/passed all automated and manual tests
Relnote: Star polygons now take two radii in the same units
Change-Id: I6444b6f660e902ec9b86e09be428c84fb489adb6