Fixed profile verifier in Android U and reenabled tests
Android U has a check to ensure the dex checksum in the profile matches
the classes.dex in the apk file. In the current version of profile
verifier tests there is a single profile used for all the tests.
As a result the dex checksum in the profile doesn't match the one of
the classes.dex for the installed apks, resulting in art service
discarding the profile when trying to compile the app with it.

Fixes: 277962514
Fixes: 276970167
Fixes: 274664672
Fixes: 277638667
Test: ./gradlew :profileinstaller:integration-tests:profile-verification:cC
Change-Id: I2487ec8160006f408456e5840b83d3e5def74af7