"Disconnect" the camera device in the previous VirtualCameraState
When 2 CameraGraphs are created and started in quick succession, it is
possible for the latter CameraGraph to create its capture session faster
than the former CameraGraph. This would cause the capture session to be
inadvertently closed later when the former CameraGraph finishes creating
its capture session.

This CL fixes this by creating a wrapper around AndroidCameraDevice
(VirtualAndroidCameraDevice), and have it be "disconnected" (reject
calls) when the VirtualCameraState is disconnected.

Bug: 269619541
Test: Run the test several times and added new unit tests
Change-Id: I888f17048b60c16c7c18982b818b0334c5380358