Temporarily add preferImmediatelyAvailableCreds to the get-passkey-option bundle
Problem: In a recent change the preferImmediatelyAvailableCreds is moved to the top request level (CreateCredentialRequest and GetCredentialRequest). However, the already-cut beta 2 OS has a version of the UI that expects the property from GetPublicKeyCredentialOption and so the new change means the sign-in flow in beta 2 will never get a true preferImmediatelyAvailableCreds value.

To prevent this regression, I will add to our jetpack library a temporary data pair to the GetPublicKeyCredentialOption when GetCredentialRequest.preferImmediatelyAvailableCreds is on, so that the UI gets this bit propagated. For the release after beta2, I will revert this change.

Bug: 278308121
Test: locally built and verified results on the UI side.
Change-Id: I6b37558b6e2d2dce5ff05bf1b39668c19d44b692