Fix issues with the KSTypeVarianceResolver.
This CL fixes issues with the KSTypeVarianceResolver, which is
responsible for returning a KSType with variance that matches the JVM
type.

This CL is essentially a complete rewrite of KSTypeVarianceResolver
since there were a number of issues that needed to be addressed
which required a different approach than was currently implemented.
In particular, this CL fixes variance for some of the following cases:

  * Usages where type parameters are resolved, e.g. via asMemberOf
  * Usages of multi-level generics with in/out variance at
    declaration/usage site.

This CL also refactors some of the existing code. For example, each of
the KspTypes used to take in a KspJvmTypeResolver which itself took in
a KspJvmTypeResolutionScope. Now, each of the KspTypes take in the
scope directly, which I think cuts out an unnecessary abstraction
layer.

A follow-up CL will fix more cases for lambdas and suspend functions.

Test: KspTypeNamesGoldenTest.kt
Change-Id: I2bc058ec9120c7d479f6bcd2c72208306b46d735