Remove duplicate links from NavAction
When NavAction was in java, we linked to different method overloads in
the Java Docs to specify which APIs NavAction should be used with.

Now that NavAction is in Kotlin, per the Kotlin docs:
"Note that KDoc does not have any syntax for resolving overloaded
members in links. Since Kotlin's documentation generation tool puts the
documentation for all overloads of a function on the same page,
identifying a specific overloaded function is not required for the link
to work." (https://kotlinlang.org/docs/kotlin-doc.html#external-links)

Test: ./gradlew bOS
Bug: 277303370
Change-Id: I172730a7cac0586f08f598c0f69150b9cb142620