Enable support for sdk 34 in profile installer
This PR enables support for sdk 34 in profile installer. Note that
the sdk in androidx repo doesn't have yet the android U constant needed
so this PR adds the value directly and we'll need to update it later
once we bump the android sdk.

Bug: 279494679
Test: ./gradlew :profileinstaller:profileinstaller:test
Change-Id: Iaf1779eb2abdd0030804240e505c68b8c39506f9
Relnote: Enabled support for Android U in profile installer